requirements.txt
RPi.GPIO
bottle
PyYAML
